+++ SKZ-ejb/src/java/dhz/skz/validatori/Validator.java
@@ -15,6 +15,6 @@ public interface Validator {
 
     public void validiraj(PodatakSirovi ps);
 
-    public void setPodaciUmjeravanja(float a, float b, float ldl, float opseg);
+    public void setPodaciUmjeravanja(Double a, Double b, Double ldl, Double opseg);
 
 }
